+++ libs/add.cpp
@@ -1,0 +1,10 @@
+//
+// Created by Levalup.
+// L.eval: Let programmer get rid of only work jobs.
+//
+
+#include "api.h"
+
+UVCXX_C_API int add(int a, int b) {
+    return a + b;
+}